+++ ledger/recovering-indexer-integration-tests/BUILD.bazel
@@ -1,0 +1,42 @@
+# Copyright (c) 2020 The DAML Authors. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+load(
+    "//bazel_tools:scala.bzl",
+    "da_scala_test_suite",
+)
+
+da_scala_test_suite(
+    name = "recovering-indexer-integration-tests",
+    srcs = glob(["src/test/suite/scala/**/*.scala"]),
+    flaky = True,
+    resources = glob(["src/test/resources/**/*"]),
+    runtime_deps = [
+        "@maven//:ch_qos_logback_logback_classic",
+        "@maven//:ch_qos_logback_logback_core",
+        "@maven//:com_h2database_h2",
+    ],
+    deps = [
+        "//daml-lf/data",
+        "//language-support/scala/bindings",
+        "//ledger/ledger-api-common",
+        "//ledger/ledger-api-domain",
+        "//ledger/ledger-api-health",
+        "//ledger/ledger-on-memory",
+        "//ledger/participant-state",
+        "//ledger/participant-state/kvutils",
+        "//ledger/sandbox:ledger-api-server",
+        "//ledger/test-common",
+        "//libs-scala/contextualized-logging",
+        "//libs-scala/resources",
+        "//libs-scala/timer-utils",
+        "@maven//:ch_qos_logback_logback_classic",
+        "@maven//:com_typesafe_akka_akka_actor_2_12",
+        "@maven//:com_typesafe_akka_akka_stream_2_12",
+        "@maven//:io_dropwizard_metrics_metrics_core",
+        "@maven//:org_mockito_mockito_core",
+        "@maven//:org_scala_lang_modules_scala_java8_compat_2_12",
+        "@maven//:org_scalactic_scalactic_2_12",
+        "@maven//:org_scalatest_scalatest_2_12",
+    ],
+)
