+++ ledger/recovering-indexer-integration-tests/src/test/resources/logback-test.xml
@@ -8,7 +8,7 @@
     </appender>
 
     <appender name="MEM" class="com.digitalasset.platform.testing.LogCollector">
-        <test>com.daml.ledger.api.server.damlonx.reference.v2.IndexerIT</test>
+        <test>com.digitalasset.platform.indexer.RecoveringIndexerIntegrationSpec</test>
     </appender>
 
     <logger name="com.digitalasset.platform.indexer.RecoveringIndexer" level="INFO">
